api/requirements.txt
Django>=3.1.9
psycopg2==2.8.6
djangorestframework
djangorestframework-simplejwt
django-cors-headers
spotipy
gunicorn
python-dotenv
tzdata